requirements.txt
chromadb
langchain
langchain_chroma
langchain_community
langchain_huggingface
langchain_openai
PyPDF2
python-dotenv
streamlit
